requirements.txt
requests
openai
python-dotenv
tweepy
apscheduler
pytz